Add test kv client
